Initial working version of memsimctl
This works on Linux and partially on OpenBSD.

OpenBSD has a quirk that needs a USB unplug replug after every command
as the FTDI driver or device gets into non responsive state. The
device still responds to every other identify command send with cu, so
it's not completely gone. If the device is powered by the target, the
memory content will remain after replug.