Fix OpenBSD
ucom (in particular ucomreadcb) inserts a \n after closing the cua
device. That \n will stay in the input buffer of the tty layer and
delivered after sending the next command and reading the response.
To get rid of any pending data in the input buffer, issue a tcflush
after opening the device.

With this change, OpenBSD works fine.